directory arrange
